+++ hp3parclient/exceptions.py
@@ -407,6 +407,10 @@ class SetQOSRuleException(SSHException):
     message = "SSH set QOS rule failed: %(command)s"
 
 
+class SrstatldException(SSHException):
+    message = "SSH command failed: %(command)s"
+
+
 class ProcessExecutionError(Exception):
     def __init__(self, stdout=None, stderr=None, exit_code=None, cmd=None,
                  description=None):