+++ docs/changelog.rst
@@ -16,6 +16,7 @@ Changes in Version 3.2.2
   was missing a description. (issue #15)
 * Add support for VLUN queries in getVLUN and getHostVLUNs when a backend
   with WSAPI 1.4.2 or greater is being used.
+* Added support for calling srstatld with a given interval and history
 
 Changes in Version 3.2.1
 ------------------------